feat: Add pagination to getAllProducts API endpoint
This commit adds pagination functionality to the `getAllProducts` API endpoint in the `productController.js` file. It introduces the `pagination` method in the `ApiFeatures` class to limit the number of products returned per page and skip products based on the current page number. The `productCount` is also included in the response to provide the total number of products available.