feat: Add user ID to createProduct API endpoint
This commit modifies the createProduct API endpoint in the productController.js file to include the user ID in the request body. By adding the user ID, the API can associate the created product with the user who made the request. This change enables better tracking and management of products created by users.