feat: Add getAllProductReviews API endpoint
This commit adds the `getAllProductReviews` API endpoint to the `productController.js` file. The endpoint allows users to retrieve all the reviews of a product by providing the product ID in the request. If the product is found, the reviews are returned in the response. This feature enhances the user experience by providing the ability to view all the reviews of a product.